refactor(Listener): Move listening to new Listener class
